refactor(ValidationAuthService): 인터페이스 생성
객체지향하도록 하기 위해 인터페이스를 만들고 상속받도록 개선했습니다.
